requirements.txt
-i https://pypi.python.org/simple
amqp==2.3.2
aniso8601==3.0.2
autopep8==1.4.3
babel==2.6.0
billiard==3.5.0.5
bleach==2.1.4
boto3==1.9.74
botocore==1.12.74
braintree==3.49.0
cairocffi==0.9.0
cairosvg==2.2.1
celery[redis]==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0 ; python_version >= '3.0'
dj-database-url==0.5.0
dj-email-url==0.1.0
django-appconf==1.0.2
django-babel==0.6.2
django-bootstrap4==0.0.7
django-cache-url==3.0.0
django-celery-results==1.0.4
django-countries==5.3.2
django-debug-toolbar-request-history==0.0.9
django-debug-toolbar==1.11
django-elasticsearch-dsl==0.5.1
django-filter==2.0.0
django-fsm==2.6.0
django-graphql-jwt==0.2.0
django-impersonate==1.3
django-js-asset==1.1.0
django-measurement==3.0.0
django-mptt==0.9.1
django-phonenumber-field==2.1.0
django-prices-openexchangerates==1.0.1
django-prices-vatlayer==1.0.2
django-prices==1.0.4
django-recaptcha==1.4.0
django-redis==4.10.0
django-render-block==0.5
django-silk==2.0.0
django-storages==1.7.1
django-templated-email==2.2.0
django-versatileimagefield==1.10
django-webpack-loader==0.6.0
django==2.1.5
docutils==0.14
elasticsearch-dsl==6.0.1
elasticsearch==6.3.1
faker==1.0.1
freezegun==0.3.11
google-i18n-address==2.3.4
google-measurement-protocol==1.0.0
gprof2dot==2016.10.13
graphene-django-optimizer==0.3.5
graphene-django==2.2.0
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
html5lib==1.0.1
idna==2.8
ipaddress==1.0.22
jinja2==2.10
jmespath==0.9.3
jsonfield==2.0.2
kombu==4.2.2.post1
markdown==3.0.1
markupsafe==1.1.0
maxminddb-geolite2==2018.703
maxminddb==1.4.1
measurement==2.0.1
mpmath==1.1.0
oauthlib==2.1.0
phonenumberslite==8.10.2
pillow==5.4.1
prices==1.0.0
promise==2.2.1
psycopg2-binary==2.7.6.1
purl==1.4
pycodestyle==2.4.0
pycparser==2.19
pygments==2.3.1
pyjwt==1.7.1
pyphen==0.9.5
python-dateutil==2.7.5 ; python_version >= '2.7'
python3-openid==3.1.0 ; python_version >= '3.0'
pytz==2018.9
raven==6.9.0
razorpay==1.1.1
redis==2.10.6
requests-oauthlib==1.0.0
requests==2.21.0
rx==1.6.1
s3transfer==0.1.13
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
sqlparse==0.2.4
stripe==2.17.0
sympy==1.3
text-unidecode==1.2
tinycss2==0.6.1
typing==3.6.6
urllib3==1.24.1
uwsgi==2.0.17.1 ; platform_system != 'Windows'
vine==1.2.0
weasyprint==44
webencodings==0.5.1
django-bootstrap-modal-forms
django-widget-tweaks
